Finished protection on alerts.php

+++ admin/alerts.php
@@ -10,27 +10,35 @@ $result = mysql_query($query) or die(mysql_error() . '<br /><br />'. $query);
 
 if (mysql_num_rows($result) > 0){
 	while($row = mysql_fetch_assoc($result)){
-		$item_ids = unserialize($row["item_ids"]);
-		$item = new item($item_ids[0]);
+		$alert = new alert($row["id"]);
+		$item = new item($alert->item_ids[0]);
 		
-		$table_rows .= '
-			<tr>
-				<td align="center" style="width: 18ex; white-space: nowrap;">
-					<nobr>
-						[<a href="edit_alert.php?id='.$row["id"].'">edit</a>]
-						[<a href="delete_alert.php?id='.$row["id"].'">delete</a>]
-					</nobr>
-				</td>
-				<td>'.$row["title"].'</td>
-				<td>'.$item->name;
-		
-		if (count($item_ids) > 1){
-			$table_rows .= ' et. al.';
-		}
+		if (is_array($alert->category_ids)){
+			foreach($alert->category_ids as $cat_id){
+				if(!$admin_user->can_admin($cat_id)){
+					break;
+				}
 				
-		$table_rows .= '
-				</td>
-			</tr>';
+				$table_rows .= '
+					<tr>
+						<td align="center" style="width: 18ex; white-space: nowrap;">
+							<nobr>
+								[<a href="edit_alert.php?id='.$row["id"].'">edit</a>]
+								[<a href="delete_alert.php?id='.$row["id"].'">delete</a>]
+							</nobr>
+						</td>
+						<td>'.$row["title"].'</td>
+						<td>'.$item->name;
+				
+				if (count($item_ids) > 1){
+					$table_rows .= ' et. al.';
+				}
+				
+				$table_rows .= '
+						</td>
+					</tr>';
+			}
+		}
 	}
 	
 	$table_rows = '<table>'.$table_rows.'</table>';
